fixed the majority of the compiler warnings; updated long line formatting

+++ tui/menu_interface.c
@@ -58,7 +58,7 @@ void themeDialog(CDKSCREEN *main_cdk_screen) {
 
         /* Check exit from widget and write the value if normal */
         if ((iface_theme_list->exitType == vNORMAL) &&
-                (g_curr_theme != theme_choice)) {
+                ((int) g_curr_theme != theme_choice)) {
             /* Load the ESOS configuration file (INI file) */
             ini_dict = iniparser_load(ESOS_CONF);
             if (ini_dict == NULL) {
